# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Siparis yonetimi icin REST API - Kafka event-driven microservice mimarisi",
        "title": "E-Commerce Order Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/orders": {
            "post": {
                "description": "Yeni bir siparis olusturur ve Kafka'ya OrderCreated eventi yayinlar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Yeni siparis olustur",
                "parameters": [
                    {
                        "description": "Siparis bilgileri",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "ID ile siparis bilgilerini getirir",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Siparis sorgula",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Siparis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.OrderStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "PAID",
                "COMPLETED",
                "CANCELLED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "OrderStatusPending",
                "OrderStatusPaid",
                "OrderStatusCompleted",
                "OrderStatusCancelled",
                "OrderStatusFailed"
            ]
        },
        "dto.CreateOrderRequest": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "string",
                    "example": "cust-001"
                },
                "product_id": {
                    "type": "string",
                    "example": "prod-001"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                },
                "total_price": {
                    "type": "number",
                    "example": 99.9
                }
            }
        },
        "dto.OrderResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "customer_id": {
                    "type": "string",
                    "example": "cust-001"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "product_id": {
                    "type": "string",
                    "example": "prod-001"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.OrderStatus"
                        }
                    ],
                    "example": "PENDING"
                },
                "total_price": {
                    "type": "number",
                    "example": 99.9
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        }
    }
}